WarWrapperTest fixed
